+++ front/src/app/page.tsx
@@ -6,10 +6,12 @@ export default function Home() {
       <h1>Il faut rouler cool!</h1>
       <h3 className="font-medium uppercase tracking-widest">Ecovoit</h3>
       <div className="flex gap-3">
+        <Button variant="outlined">outlined</Button>
         <Button variant="outlined" color="secondary">
           outlined
         </Button>
-        <Button color="primary">Normal</Button>
+        <Button>text</Button>
+        <Button color="secondary">text</Button>
         <Button variant={"contained"}>Contained</Button>
         <Button variant={"contained"} color="secondary">
           Contained
